+++ Frameworks/GnuSpeech/Parsers/TextParser/NXStream.m
@@ -170,7 +170,7 @@
 		return;
 	}
 	
-	int buflen = strlen(buf);
+	long buflen = strlen(buf);
 
 	NSRange range;
 	range.location = streamPosition;
@@ -182,7 +182,7 @@
 		
 	} else {
 		
-		int streamBufferLength = [streamBuffer length];
+		NSUInteger streamBufferLength = [streamBuffer length];
 		if (streamPosition > (int)streamBufferLength - (int)range.length) {  // range to write is beyond string bounds; note necessary casts to (int)
 		
 			range.length = streamBufferLength - streamPosition;  // adjust range to be within stream buffer bounds
